+++ blockchain/forkcontroller_test.go
@@ -39,23 +39,23 @@ func TestForkController(t *testing.T) {
 	}
 	fc := NewForkController(testcfg, params.DefaultChainconfig)
 
-	var height int64
+	var number int64
 	for j := 0; j < 2; j++ {
 		for i := 0; i < 8; i++ {
 
-			block := &types.Block{Head: &types.Header{Number: big.NewInt(height)}}
+			block := &types.Block{Head: &types.Header{Number: big.NewInt(number)}}
 			block.Head.WithForkID(uint64(j), uint64(j+1))
 			assert.NoError(t, fc.checkForkID(block.Header(), statedb))
 			assert.NoError(t, fc.update(block, statedb))
-			height++
+			number++
 		}
 
 		for i := 0; i < 10; i++ {
-			block := &types.Block{Head: &types.Header{Number: big.NewInt(height)}}
+			block := &types.Block{Head: &types.Header{Number: big.NewInt(number)}}
 			block.Head.WithForkID(uint64(j+1), uint64(j+1))
 			assert.NoError(t, fc.checkForkID(block.Header(), statedb))
 			assert.NoError(t, fc.update(block, statedb))
-			height++
+			number++
 		}
 
 		id, _, err := fc.currentForkID(statedb)
@@ -77,14 +77,14 @@ func TestUpdateDifferentForkBlock(t *testing.T) {
 	}
 
 	fc := NewForkController(testcfg, params.DefaultChainconfig)
-	var height int64
+	var number int64
 	for j := 0; j < 2; j++ {
 		for i := 0; i < 7; i++ {
-			block := &types.Block{Head: &types.Header{Number: big.NewInt(height)}}
+			block := &types.Block{Head: &types.Header{Number: big.NewInt(number)}}
 			block.Head.WithForkID(uint64(0), uint64(j+1))
 			assert.NoError(t, fc.checkForkID(block.Header(), statedb))
 			assert.NoError(t, fc.update(block, statedb))
-			height++
+			number++
 
 			info, err := fc.getForkInfo(statedb)
 			if err != nil {