+++ processor/processor.go
@@ -106,7 +106,6 @@ func (p *StateProcessor) ApplyTransaction(author *common.Name, gp *common.GasPoo
 	detailTx := &types.DetailTx{}
 	var detailActions []*types.DetailAction
 	for i, action := range tx.GetActions() {
-		//
 		if needCheckSign(accountDB, action) {
 			if err := accountDB.RecoverTx(types.NewSigner(config.ChainID), tx); err != nil {
 				return nil, 0, err