+++ rpcapi/txpool.go
@@ -24,18 +24,18 @@ import (
 	"github.com/fractalplatform/fractal/types"
 )
 
-// PublicTxPoolAPI offers and API for the transaction pool. It only operates on data that is non confidential.
-type PublicTxPoolAPI struct {
+// PrivateTxPoolAPI offers and API for the transaction pool. It only operates on data that is non confidential.
+type PrivateTxPoolAPI struct {
 	b Backend
 }
 
-// NewPublicTxPoolAPI creates a new tx pool service that gives information about the transaction pool.
-func NewPublicTxPoolAPI(b Backend) *PublicTxPoolAPI {
-	return &PublicTxPoolAPI{b}
+// NewPrivateTxPoolAPI creates a new tx pool service that gives information about the transaction pool.
+func NewPrivateTxPoolAPI(b Backend) *PrivateTxPoolAPI {
+	return &PrivateTxPoolAPI{b}
 }
 
 // Status returns the number of pending and queued transaction in the pool.
-func (s *PublicTxPoolAPI) Status() map[string]int {
+func (s *PrivateTxPoolAPI) Status() map[string]int {
 	pending, queue := s.b.Stats()
 	return map[string]int{
 		"pending": pending,
@@ -44,13 +44,12 @@ func (s *PublicTxPoolAPI) Status() map[string]int {
 }
 
 // Content returns the transactions contained within the transaction pool.
-func (s *PublicTxPoolAPI) Content() map[string]map[string]map[string]*types.RPCTransaction {
+func (s *PrivateTxPoolAPI) Content() map[string]map[string]map[string]*types.RPCTransaction {
 	content := map[string]map[string]map[string]*types.RPCTransaction{
 		"pending": make(map[string]map[string]*types.RPCTransaction),
 		"queued":  make(map[string]map[string]*types.RPCTransaction),
 	}
 	pending, queue := s.b.TxPoolContent()
-
 	// Flatten the pending transactions
 	for account, txs := range pending {
 		dump := make(map[string]*types.RPCTransaction)
@@ -70,7 +69,7 @@ func (s *PublicTxPoolAPI) Content() map[string]map[string]map[string]*types.RPCT
 	return content
 }
 
-// SetGasPrice set gas price limit
-func (s *PublicTxPoolAPI) SetGasPrice(gasprice *big.Int) bool {
-	return s.b.SetGasPrice(gasprice)
+// SetGasPrice set gas price
+func (s *PrivateTxPoolAPI) SetGasPrice(gasprice uint64) bool {
+	return s.b.SetGasPrice(big.NewInt(int64(gasprice)))
 }